fix: change manageable tokens to be a derived of network tokens
