+++ src/frontend/src/lib/derived/network.derived.ts
@@ -1,4 +1,6 @@
 import { icrcAccountIdentifierText } from '$icp/derived/ic.derived';
+import { buildIcrcCustomTokens } from '$icp/services/icrc-custom-tokens.services';
+import type { IcrcCustomToken } from '$icp/types/icrc-custom-token';
 import { DEFAULT_NETWORK, DEFAULT_NETWORK_ID } from '$lib/constants/networks.constants';
 import { address } from '$lib/derived/address.derived';
 import { routeNetwork } from '$lib/derived/nav.derived';
@@ -6,12 +8,13 @@ import { networks } from '$lib/derived/networks.derived';
 import { tokens } from '$lib/derived/tokens.derived';
 import type { OptionAddress } from '$lib/types/address';
 import type { Network, NetworkId } from '$lib/types/network';
-import type { Token } from '$lib/types/token';
+import type { ManageableToken, Token } from '$lib/types/token';
 import {
 	isNetworkIdChainFusion,
 	isNetworkIdEthereum,
 	isNetworkIdICP
 } from '$lib/utils/network.utils';
+import { mergeTokenLists } from '$lib/utils/token.utils';
 import { nonNullish } from '@dfinity/utils';
 import { derived, type Readable } from 'svelte/store';
 
@@ -50,4 +53,21 @@ export const networkAddress: Readable<OptionAddress | string> = derived(
 	[address, icrcAccountIdentifierText, networkICP],
 	([$address, $icrcAccountIdentifierStore, $networkICP]) =>
 		$networkICP ? $icrcAccountIdentifierStore : $address
+);
+
+export const manageableNetworkTokens: Readable<ManageableToken[]> = derived(
+	[networkTokens],
+	([$networkTokens]) => {
+		const allIcrcCustomTokens: IcrcCustomToken[] = buildIcrcCustomTokens()
+			.map((token) => ({
+				...token,
+				id: Symbol(token.symbol),
+				enabled: false
+			}))
+			.filter((token) => token.indexCanisterVersion !== 'outdated');
+		return mergeTokenLists<ManageableToken>(
+			$networkTokens.map((token) => ({ ...token, enabled: true })),
+			allIcrcCustomTokens
+		);
+	}
 );
