+++ src/frontend/src/lib/derived/tokens.derived.ts
@@ -2,10 +2,7 @@ import { ICP_TOKEN } from '$env/tokens.env';
 import { erc20Tokens } from '$eth/derived/erc20.derived';
 import { enabledEthereumTokens } from '$eth/derived/tokens.derived';
 import { sortedIcrcTokens } from '$icp/derived/icrc.derived';
-import { buildIcrcCustomTokens } from '$icp/services/icrc-custom-tokens.services';
-import type { IcrcCustomToken } from '$icp/types/icrc-custom-token';
-import type { ManageableToken, Token } from '$lib/types/token';
-import { mergeTokenLists } from '$lib/utils/token.utils';
+import type { Token } from '$lib/types/token';
 import { derived, type Readable } from 'svelte/store';
 
 export const tokens: Readable<Token[]> = derived(
@@ -17,17 +14,3 @@ export const tokens: Readable<Token[]> = derived(
 		...$erc20Tokens
 	]
 );
-
-export const manageableTokens: Readable<ManageableToken[]> = derived([tokens], ([$tokens]) => {
-	const allIcrcCustomTokens: IcrcCustomToken[] = buildIcrcCustomTokens()
-		.map((token) => ({
-			...token,
-			id: Symbol(token.symbol),
-			enabled: false
-		}))
-		.filter((token) => token.indexCanisterVersion !== 'outdated');
-	return mergeTokenLists<ManageableToken>(
-		$tokens.map((token) => ({ ...token, enabled: true })),
-		allIcrcCustomTokens
-	);
-});
